refactor: move component and styling files to new directories
moves CloudWatchContainer, ConfigPageComponent, ConfigPageContainer, ColdStartsMetricsContainer components and style sheets to new directories
